+++ packages/client/src/scripts/get-user-menu.ts
@@ -256,6 +256,27 @@ export function getUserMenu(user, router: Router = mainRouter) {
 			},
 		},
 		{
+			icon: `${icon("ph-share")}`,
+			text: i18n.ts.share,
+			type: "parent",
+			children: [
+				{
+					icon: "ph-qr-code ph-bold ph-lg",
+					text: i18n.ts.getQrCode,
+					action: () => {
+						os.displayQrCode(`https://${host}/follow-me?acct=${user.username}`);
+					},
+				},
+				{
+					icon: `${icon("ph-hand-waving")}`,
+					text: i18n.ts.copyRemoteFollowUrl,
+					action: () => {
+						copyToClipboard(`https://${host}/follow-me?acct=${user.username}`);
+					},
+				},
+			],
+		},
+		{
 			icon: `${icon("ph-newspaper")}`,
 			text: i18n.ts._feeds.copyFeed,
 			type: "parent",
@@ -290,6 +311,15 @@ export function getUserMenu(user, router: Router = mainRouter) {
 				os.post({ specified: user });
 			},
 		},
+		!isSignedIn(me)
+			? {
+					icon: `${icon("ph-hand-waving")}`,
+					text: i18n.ts.remoteFollow,
+					action: () => {
+						router.push(`/follow-me?acct=${user.username}`);
+					},
+				}
+			: undefined,
 		isSignedIn(me) && me.id !== user.id
 			? {
 					type: "link",